download

+++ package/chemicalchecker/core/neig.py
@@ -99,9 +99,9 @@ class neig(BaseSignature, DataSignature):
                 dh5out.create_dataset("row_keys", data=dh5["keys"][:])
                 dh5out["col_keys"] = h5py.SoftLink('/row_keys')
                 dh5out.create_dataset(
-                    "indices", (self.datasize[0], k), dtype=int )
+                    "indices", (self.datasize[0], k), dtype='int32' )
                 dh5out.create_dataset(
-                    "distances", (self.datasize[0], k), dtype=float)
+                    "distances", (self.datasize[0], k), dtype='float32')
                 dh5out.create_dataset("shape", data=(self.datasize[0], k))
                 dh5out.create_dataset(
                     "date", data=[datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S").encode(encoding='UTF-8', errors='strict')])
@@ -114,12 +114,12 @@ class neig(BaseSignature, DataSignature):
                     index = faiss.IndexFlatIP(self.datasize[1])
 
                 for chunk in sign.chunker():
-                    data_temp = np.array(dh5["V"][chunk], dtype=float)
+                    data_temp = np.array(dh5["V"][chunk], dtype='float32')
                     if self.metric == "cosine":
                         normst = LA.norm(data_temp, axis=1)
                         index.add( np.array( data_temp / normst[:, None], dtype='float32') )
                     else:
-                        index.add( np.array(data_temp, dtype='float32') )
+                        index.add( np.array(data_temp, dtype='float32')  )
 
                 for chunk in sign.chunker():
                     data_temp = np.array(dh5["V"][chunk], dtype=float)
@@ -176,9 +176,9 @@ class neig(BaseSignature, DataSignature):
                 with h5py.File(self.data_path, 'r') as hr5:
                     dh5out.create_dataset("col_keys", data=hr5["row_keys"][:])
                 dh5out.create_dataset(
-                    "indices", (self.datasize[0], k), dtype=int )
+                    "indices", (self.datasize[0], k), dtype='int32' )
                 dh5out.create_dataset(
-                    "distances", (self.datasize[0], k), dtype=float)
+                    "distances", (self.datasize[0], k), dtype='float32')
                 dh5out.create_dataset("shape", data=(self.datasize[0], k))
                 dh5out.create_dataset(
                     "date", data=[datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S").encode(encoding='UTF-8', errors='strict')])
@@ -406,7 +406,7 @@ class neig(BaseSignature, DataSignature):
             self.__log.warning("Maximum k is %s.", max_k)
             k = max_k
         # convert signatures to float32 as faiss is very picky
-        data = np.array(signatures, dtype=float)
+        data = np.array(signatures, dtype='float32')
         self.__log.info("Searching %s neighbors" % k)
         # get neighbors idx and distances
         if "cosine" in metric_orig:
